Graph widget: mokey-Patch SVGCanvas to not call moveTo() on beginPath()
Due to the original behavior of beginPath(), each path was started with
a moveTo(0,0) command. This introduced problems when arrow heads were
exported through SVGCavas by Cytoscape.js. Instead of patching the
resulting SVG, it is easier to do actually don't do the moveTo() at the
start of a new path.